refactor(chunk-store): Phase 2b — move refs bodies to chunk_refs.c
Moves all refs-touching function bodies (~305 lines) from
chunk_store.c to chunk_refs.c, completing the refs subsystem
extraction begun in #875.

Functions moved:
  Already non-static (Phase 1b exposed; bodies now relocated):
    csCaptureSavedRefsState, csRestoreSavedRefsState,
    csFreeSavedRefsState, csFreeRefsState,
    csEnsureDefaultBranch, csAdoptRefsState,
    csReplaceRefsStateFromBlob

  Newly exposed in chunk_refs.h (called from chunk_store.c):
    csFreeBranches, csFreeTags, csFreeRemotes, csFreeTracking
    csMarkRefsCommitted, csRestoreCommittedRefsHash
    csDetachSavedRefsState
    csDeserializeRefs, csDeserializeRefsIntoTemp

chunk_refs.c now #includes chunk_store.h to access the
ChunkStore struct fields, and locally defines CS_READ_U32/
CS_READ_I64 (same pattern chunk_wal.c uses).

This is the third and final "phase Xb" of the chunk-store
decomposition — combined with the merged Phase 1b (#890) and
the open Phase 4b (#892) and 5b (#891), chunk_store.c becomes a
thin orchestration layer over the chunk_wal, chunk_refs,
chunk_index, chunk_staging, and chunk_file subsystems.

Verified:
  - make doltlite-lib + make doltlite (clean builds)
  - make lint (layering check passes)
  - Smoke: CREATE + dolt_commit + reopen + dolt_log count = 2;
    dolt_branch('feat') visible in dolt_branches.
  - WAL regression tests (37/37 pass)
  - Full C regression suite: 108622/108622 pass
